requirements.txt
agileutil==12.85
aiofiles==0.4.0
astroid==2.2.2
autopep8==1.4.3
certifi==2019.11.28
chardet==3.0.4
cheroot==8.2.1
DBUtils==1.3
demjson==2.2.4
gunicorn==20.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
isort==4.3.10
jaraco.functools==2.0
Jinja2==2.10.3
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.0.0
multidict==4.6.1
pexpect==4.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pylint==2.3.1
PyMySQL==0.9.3
python-decouple==3.1
python-ldap==3.1.0
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
sanic-session==0.5.6
six==1.13.0
typed-ast==1.2.0
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
uWSGI==2.0.18
web.py==0.40.dev1
websockets==8.1
wrapt==1.11.1
yapf==0.27.0
sanic==19.9.0
sanic-session==0.5.6